CDEC-433: Push network related code from block package up into lib
