+++ block/src/Pos/Block/Types.hs
@@ -8,15 +8,11 @@ module Pos.Block.Types
        , LastKnownHeader
        , LastKnownHeaderTag
        , MonadLastKnownHeader
-
-       , RecoveryHeaderTag
-       , RecoveryHeader
-       , MonadRecoveryHeader
        ) where
 
 import           Universum
 
-import qualified Control.Concurrent.STM as STM
+import qualified Data.Text.Buildable
 import           Formatting (bprint, build, (%))
 import qualified Formatting.Buildable
 import           Serokell.Util.Text (listJson)
@@ -28,7 +24,6 @@ import           Pos.Core (HasConfiguration, HasDifficulty (..),
 import           Pos.Core.Block (Block, BlockHeader)
 import           Pos.Core.Txp (TxpUndo)
 import           Pos.Delegation.Types (DlgUndo)
-import           Pos.Infra.Communication.Protocol (NodeId)
 import           Pos.Update.Poll.Types (USUndo)
 import           Pos.Util.Util (HasLens (..))
 
@@ -66,11 +61,6 @@ data LastKnownHeaderTag
 type LastKnownHeader = TVar (Maybe BlockHeader)
 type MonadLastKnownHeader ctx m
      = (MonadReader ctx m, HasLens LastKnownHeaderTag ctx LastKnownHeader)
-
-data RecoveryHeaderTag
-type RecoveryHeader = STM.TMVar (NodeId, BlockHeader)
-type MonadRecoveryHeader ctx m
-     = (MonadReader ctx m, HasLens RecoveryHeaderTag ctx RecoveryHeader)
 
 -- TH derived instances at the end of the file.
