+++ lib/src/Pos/Context/Context.hs
@@ -26,12 +26,9 @@ import           Control.Lens (lens, makeLensesWith)
 import           Data.Time.Clock (UTCTime)
 import           System.Wlog (LoggerConfig)
 
-import           Pos.Block.RetrievalQueue (BlockRetrievalQueue,
-                     BlockRetrievalQueueTag)
 import           Pos.Block.Slog (HasSlogContext (..), HasSlogGState (..),
                      SlogContext (..))
-import           Pos.Block.Types (LastKnownHeader, LastKnownHeaderTag,
-                     RecoveryHeader, RecoveryHeaderTag)
+import           Pos.Block.Types (LastKnownHeader, LastKnownHeaderTag)
 import           Pos.Communication.Types (NodeId)
 import           Pos.Core (HasPrimaryKey (..), Timestamp)
 import           Pos.Infra.DHT.Real.Param (KademliaParams)
@@ -47,6 +44,9 @@ import           Pos.Infra.StateLock (StateLock, StateLockMetrics)
 import           Pos.Infra.Util.JsonLog.Events (MemPoolModifyReason (..))
 import           Pos.Launcher.Param (BaseParams (..), NodeParams (..))
 import           Pos.Lrc.Context (LrcContext)
+import           Pos.Network.Block.RetrievalQueue (BlockRetrievalQueue,
+                     BlockRetrievalQueueTag)
+import           Pos.Recovery.Types (RecoveryHeader, RecoveryHeaderTag)
 import           Pos.Ssc.Types (HasSscContext (..), SscContext)
 import           Pos.Txp.Settings (TxpGlobalSettings)
 import           Pos.Update.Context (UpdateContext)
