+++ lib/src/Pos/WorkMode/Class.hs
@@ -21,7 +21,7 @@ import           UnliftIO (MonadUnliftIO)
 import           Pos.Block.BListener (MonadBListener)
 import           Pos.Block.Configuration (HasBlockConfiguration)
 import           Pos.Block.Slog (HasSlogContext, HasSlogGState)
-import           Pos.Block.Types (MonadLastKnownHeader, MonadRecoveryHeader)
+import           Pos.Block.Types (MonadLastKnownHeader)
 import           Pos.Configuration (HasNodeConfiguration)
 import           Pos.Context (BlockRetrievalQueue, BlockRetrievalQueueTag,
                      HasSscContext, StartTime, TxpGlobalSettings)
@@ -40,6 +40,7 @@ import           Pos.Infra.StateLock (StateLock, StateLockMetrics)
 import           Pos.Infra.Util.JsonLog.Events (MemPoolModifyReason)
 import           Pos.Infra.Util.TimeWarp (CanJsonLog)
 import           Pos.Lrc.Context (HasLrcContext)
+import           Pos.Recovery.Types (MonadRecoveryHeader)
 import           Pos.Security.Params (SecurityParams)
 import           Pos.Ssc (HasSscConfiguration)
 import           Pos.Ssc.Mem (MonadSscMem)
